+++ lms/djangoapps/commerce/api/v1/serializers.py
@@ -19,7 +19,7 @@ from .models import UNDEFINED, Course
 class CourseModeSerializer(serializers.ModelSerializer):
     """ CourseMode serializer. """
     name = serializers.CharField(source='mode_slug')
-    price = serializers.IntegerField(source='min_price')
+    price = serializers.DecimalField(source='min_price', max_digits=30, decimal_places=2)
     expires = serializers.DateTimeField(
         source='expiration_datetime',
         required=False,
